requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
iniconfig==2.0.0
mock==5.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
mysqlx-connector-python==9.1.0
packaging==24.2
pluggy==1.5.0
protobuf==4.25.3
psycopg2-binary==2.9.10
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
sqlparse==0.5.3